Criação da query para o relatório Relatorio Qualificacao Recusa Familiar

+++ src/main/java/br/ifes/leds/sincap/gerenciaNotificacao/cln/cdp/relatorios/QualificacaoRecusaFamiliar.java
@@ -15,5 +15,5 @@ import lombok.experimental.Builder;
 public class QualificacaoRecusaFamiliar {
     private Long codigo;
     private String nomeRecusa;
-    private Double total;
+    private Long total;
 }
